Forums: Overzealous type coercion caused `bbp_forum_query_last_reply_id()` to skip the database query in every instance (since 0 is never === false). As such, the resetting of `_bbp_last_reply_id` got short-circuted in some cases.
This fix is for the 2.5 branch, `/trunk` fixed this inadvertantly in [5954]

Props boonebgorges, mauteri.
Fixes #2414.


git-svn-id: https://bbpress.svn.wordpress.org/branches/2.5@6020 9866e705-20ec-0310-96e7-cbb4277adcfb

+++ includes/forums/functions.php
@@ -1931,7 +1931,7 @@ function bbp_forum_query_last_reply_id( $forum_id, $topic_ids = 0 ) {
 	global $wpdb;
 
 	$cache_id = 'bbp_get_forum_' . $forum_id . '_reply_id';
-	$reply_id = (int) wp_cache_get( $cache_id, 'bbpress_posts' );
+	$reply_id = wp_cache_get( $cache_id, 'bbpress_posts' );
 
 	if ( false === $reply_id ) {
